Design By Numbers 1.0.1 archive
This is an archived copy of the DBN 1.0.1 project.

Repo was created from a backup copy of the project
directory years later (2018). All files are unchanged
with their original timestamps.

Note this was essentially an internal release
Tom White wrote for John Maeda as a first attempt
to implement Maeda's DBN language spec.